+++ packages/eyeglass/src/importers/ImportUtilities.ts
@@ -1,6 +1,6 @@
 import * as debug from "../util/debug";
 import { URI } from "../util/URI";
-import merge = require("lodash.merge");
+import merge from "lodash.merge";
 import { ImporterReturnType, AsyncImporter, AsyncContext } from "node-sass";
 import { IEyeglass } from "../IEyeglass";
 import { SassImplementation } from "../util/SassImplementation";